research/cross-layer-svd/requirements.txt
torch>=2.2
transformers>=4.44
datasets>=2.19
numpy
tqdm
python-dotenv
httpx
